Add refresh button and 2-row nav to overview page
